requirements.txt
defusedxml==0.5.0
Django==1.11
django-el-pagination==3.1.0
gunicorn==19.7.1
oauthlib==2.0.2
psycopg2==2.7.1
PyJWT==1.5.0
python3-openid==3.1.0
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
whitenoise==3.3.0
djangorestframework==3.9.0
-e git+git@github.com:fcornelius/django-firebase-auth.git#egg=firebase_auth